docs(changelog): expand Test Lab decomposition entry with pass-2 details
